CNN Brain Tumor Classification
The main purpose of this project was to build a CNN model that would classify if subject has a tumor or not based on MRI scan. I used the VGG-16, VGG-19, Inception v3 model architecture and custom built model to train the model for this binary problem. Used accuracy as a metric to justify the model performance.